api/requirements.txt
aniso8601==7.0.0
astroid==2.2.5
autopep8==1.4.4
Click==7.0
Flask==1.1.1
flask-marshmallow==0.10.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
marshmallow==3.0.0
marshmallow-sqlalchemy==0.17.0
mccabe==0.6.1
psycopg2==2.8.3
pycodestyle==2.5.0
pylint==2.3.1
python-dotenv==0.10.3
pytz==2019.2
six==1.12.0
SQLAlchemy==1.3.7
Werkzeug==0.15.5
wrapt==1.11.2
